// src/app/core/enums/sys/permission-values.enum.ts
export enum PermissionValues {
  DASHBOARD_ADMIN = 'DASHBOARD_ADMIN',
  DASHBOARD_DOCTOR = 'DASHBOARD_DOCTOR',
  DASHBOARD_RECEPTIONIST = 'DASHBOARD_RECEPTIONIST',
  MOD_CLINICA = 'MOD_CLINICA',
  GESTION_DIAGNOSTICO = 'GESTION_DIAGNOSTICO',
  GESTION_CONSULTORIO = 'GESTION_CONSULTORIO',
  GESTION_PACIENTES = 'GESTION_PACIENTES',
  GESTION_MEDICOS = 'GESTION_MEDICOS',
  GESTION_AFECCIONES_MED = 'GESTION_AFECCIONES_MED',
  GESTION_ESPECIALIDADES = 'GESTION_ESPECIALIDADES',
  MOD_CONFIGURACIONES = 'MOD_CONFIGURACIONES',
  GESTION_CONF_CLIENTE = 'GESTION_CONF_CLIENTE',
  GESTION_TIPO_DOCS = 'GESTION_TIPO_DOCS',
  MOD_PROG_MEDICO = 'MOD_PROG_MEDICO',
  GESTION_PROG_MEDIC = 'GESTION_PROG_MEDIC',
  MOD_VENTAS = 'MOD_VENTAS',
  FORM_VENTA = 'FORM_VENTA',
  HISTORIAL_VENTAS = 'HISTORIAL_VENTAS',
  MOD_FINANZAS = 'MOD_FINANZAS',
  GESTION_PERIDOS_CONT = 'GESTION_PERIDOS_CONT',
  GESTION_MOV_CONTABLES = 'GESTION_MOV_CONTABLES',
  GESTION_MONEDAS = 'GESTION_MONEDAS',
  MOD_PRODUCTOS = 'MOD_PRODUCTOS',
  GESTION_PRODUCTOS = 'GESTION_PRODUCTOS',
  GESTION_TIPO_PRODUCTOS = 'GESTION_TIPO_PRODUCTOS',
  GESTION_COND_PAGOS = 'GESTION_COND_PAGOS',
  MOD_SEGURIDAD = 'MOD_SEGURIDAD',
  GESTION_USUARIOS = 'GESTION_USUARIOS',
  GESTION_ROLES = 'GESTION_ROLES',
  GESTION_PERMISOS = 'GESTION_PERMISOS',
  GESTION_RECURSOS = 'GESTION_RECURSOS',
  MOD_RRHH = 'MOD_RRHH',
  GESTION_EMPLEADOS = 'GESTION_EMPLEADOS',
  ABM_MEDICAMENTOS = 'ABM_MEDICAMENTOS',
}